requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
openpyxl
streamlit
psycopg2-binary
pgvector
langchain
langchain-ollama
langchain_openai
ollama
selenium
webdriver-manager
transformers
torch
langchain_community
fitz
frontend
python-pptx
python-docx
google-genai
pycirclize
-e .
